add offer-page-component
